+++ coloraide/distance/__init__.py
@@ -45,17 +45,17 @@ def distance_euclidean(color: Color, sample: Color, space: str = "lab-d65") -> f
     # convert to the specified space
     c1 = color.convert(space, norm=False)
     c2 = sample.convert(space, norm=False)
-    coords1 = c1.coords()
-    coords2 = c2.coords()
+    coords1 = c1.coords(nans=False)
+    coords2 = c2.coords(nans=False)
 
     # Convert polar coordinate into rectangular coordinates
     if c1.is_polar():
         hi = c1._space.hue_index()  # type: ignore[attr-defined]
         ri = c1._space.radial_index()  # type: ignore[attr-defined]
-        a, b = alg.polar_to_rect(c1[ri], c1[hi])
+        a, b = alg.polar_to_rect(coords1[ri], coords1[hi])
         coords1[hi] = a
         coords1[ri] = b
-        a, b = alg.polar_to_rect(c2[ri], c2[hi])
+        a, b = alg.polar_to_rect(coords2[ri], coords2[hi])
         coords2[hi] = a
         coords2[ri] = b