+++ coloraide/color.py
@@ -536,7 +536,7 @@ class Color(metaclass=ColorMeta):
         """Normalize the color."""
 
         self[:-1] = self._space.normalize(self.coords(nans=False))
-        if nans and self.is_polar() and self.is_achromatic():
+        if nans and self._space.is_polar() and self.is_achromatic():
             i = self._space.hue_index()  # type: ignore[attr-defined]
             self[i] = math.nan
         alpha = self[-1]
@@ -607,7 +607,7 @@ class Color(metaclass=ColorMeta):
         this._coords[:-1] = coords
 
         # Normalize achromatic colors, but skip if we internally don't need this.
-        if norm and this.is_polar() and this.is_achromatic():
+        if norm and this._space.is_polar() and this.is_achromatic():
             this[this._space.hue_index()] = math.nan  # type: ignore[attr-defined]
 
         return this
@@ -1350,11 +1350,6 @@ class Color(metaclass=ColorMeta):
             return self[-1]
         else:
             return self._space.resolve_channel(-1, self._coords)
-
-    def is_polar(self) -> bool:
-        """Check if space is polar."""
-
-        return self._space.is_polar()
 
 
 Color.register(