+++ coloraide/distance/__init__.py
@@ -49,7 +49,7 @@ def distance_euclidean(color: Color, sample: Color, space: str = "lab-d65") -> f
     coords2 = c2.coords(nans=False)
 
     # Convert polar coordinate into rectangular coordinates
-    if c1.is_polar():
+    if c1._space.is_polar():
         hi = c1._space.hue_index()  # type: ignore[attr-defined]
         ri = c1._space.radial_index()  # type: ignore[attr-defined]
         a, b = alg.polar_to_rect(coords1[ri], coords1[hi])